+++ src/app/dashboard/beers/page.tsx
@@ -1,4 +1,4 @@
-import { BeerList } from "@/components/dashboard/beer-list";
+import { BeerList } from "@/components/dashboard/beers/beer-list";
 import { hasAccessToBrewery } from "@/lib/auth/access";
 import { getBreweryBeers } from "@/lib/breweries";
 import { notFound } from "next/navigation";
