+++ src/app/dashboard/profile/page.tsx
@@ -1,5 +1,5 @@
-import { ProfileImageUpload } from "@/components/dashboard/profile-image-upload";
-import { SignOut } from "@/components/dashboard/sign-out";
+import { ProfileImageUpload } from "@/components/dashboard/profile/profile-image-upload";
+import { SignOut } from "@/components/dashboard/profile/sign-out";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
