+++ src/app/dashboard/taps/page.tsx
@@ -1,6 +1,6 @@
 import { AddTap } from "@/components/add-tap";
-import { BeerDetail } from "@/components/dashboard/beer-detail";
-import { DeleteTapButton } from "@/components/dashboard/delete-tap-button";
+import { BeerDetail } from "@/components/dashboard/beers/beer-detail";
+import { DeleteTapButton } from "@/components/dashboard/taps/delete-tap-button";
 import {
   Card,
   CardContent,
