+++ src/components/dashboard/bar/bar-details-form.tsx
