+++ src/components/dashboard/bar/bar-details.tsx
