+++ src/components/dashboard/beers/beer-detail.tsx
