+++ src/components/dashboard/beers/delete-beer-dialog.tsx
