+++ src/components/dashboard/taps/delete-tap-button.tsx
